refactor(transfer): split receiver/transfer.rs into per-concern submodules
Decomposes the 1058-line receiver transfer driver into focused submodules
under `crates/transfer/src/receiver/transfer/`:

- `sync.rs` - synchronous per-file loop, with the delta-token apply path
  factored out as a free `apply_delta_tokens` helper plus a sparse-aware
  `write_chunk` shim.
- `pipelined.rs` - two-phase pipelined driver (phase 1 short-sum, phase 2
  redo at full sum length).
- `pipelined_incremental.rs` - same as `pipelined` plus incremental
  directory creation and failed-dir tracking.
- `setup.rs` - shared `setup_transfer` plus the file-private
  `parse_wire_filters_for_receiver` filter-rule converter.

The parent `transfer.rs` retains only the `run` dispatcher and the new
submodule declarations. Public API is unchanged: every method that lived
on `ReceiverContext` keeps its prior path and visibility, and the private
`PipelineSetup`, `PHASE1_CHECKSUM_LENGTH`, and `REDO_CHECKSUM_LENGTH`
items in `receiver/mod.rs` are reached as descendant items from the new
submodules without elevating their visibility. `setup_transfer` widens
from `pub(super)` to `pub(in crate::receiver)` so the rustdoc link in
`receiver/mod.rs` keeps resolving.

All submodules land well under the 650 LoC cap (largest is `sync.rs` at
505 lines).
