+++ crates/transfer/src/receiver/transfer/setup.rs
@@ -1,0 +1,245 @@
+//! Common transfer setup shared by every receiver entry point.
+//!
+//! `setup_transfer` activates input multiplex, reads the filter list, receives
+//! the (possibly incremental) file list, sanitizes paths, and builds the
+//! `PipelineSetup` used by `run_sync`, `run_pipelined`, and
+//! `run_pipelined_incremental`. Filter-list wire parsing lives alongside it in
+//! `parse_wire_filters_for_receiver`.
+
+use std::io::{self, Read};
+use std::path::PathBuf;
+use std::sync::Arc;
+
+use logging::{debug_log, info_log};
+use metadata::MetadataOptions;
+use protocol::filters::{FilterRuleWireFormat, RuleType, read_filter_list};
+
+use filters::{DirMergeConfig, FilterChain, FilterSet};
+
+use crate::receiver::{PHASE1_CHECKSUM_LENGTH, PipelineSetup, ReceiverContext};
+use crate::shared::ChecksumFactory;
+
+impl ReceiverContext {
+    /// Common setup for all transfer modes.
+    ///
+    /// Activates input multiplex, reads filter list if needed, receives the file
+    /// list (including INC_RECURSE extra segments), sanitizes paths, and builds
+    /// the `PipelineSetup` with checksum and metadata configuration.
+    ///
+    /// # Upstream Reference
+    ///
+    /// - `main.c:1342-1343` - client receiver activates multiplex at protocol >= 23
+    /// - `main.c:1167-1168` - server receiver activates multiplex at protocol >= 30
+    pub(in crate::receiver) fn setup_transfer<R: Read>(
+        &mut self,
+        reader: crate::reader::ServerReader<R>,
+    ) -> io::Result<(crate::reader::ServerReader<R>, usize, PipelineSetup)> {
+        // upstream: generator.c:2260-2261 - emitted at the top of generate_files,
+        // just before the per-segment dispatch loop. The receiver-side transfer
+        // setup is the closest analog (every `run*` entry point routes through
+        // setup_transfer).
+        debug_log!(Genr, 1, "generator starting pid={}", std::process::id());
+
+        let mut reader = if self.should_activate_input_multiplex() {
+            reader.activate_multiplex().map_err(|e| {
+                io::Error::new(
+                    e.kind(),
+                    format!(
+                        "failed to activate INPUT multiplex: {e} {}{}",
+                        crate::role_trailer::error_location!(),
+                        crate::role_trailer::receiver()
+                    ),
+                )
+            })?
+        } else {
+            reader
+        };
+
+        if self.should_read_filter_list() {
+            let wire_rules = read_filter_list(&mut reader, self.protocol).map_err(|e| {
+                io::Error::new(
+                    e.kind(),
+                    format!(
+                        "failed to read filter list: {e} {}{}",
+                        crate::role_trailer::error_location!(),
+                        crate::role_trailer::receiver()
+                    ),
+                )
+            })?;
+
+            // upstream: clientserver.c:rsync_module() - daemon_filter_list is applied
+            // on top of client filters. Daemon rules take precedence (prepended).
+            let daemon_rules = &self.config.daemon_filter_rules;
+            let combined = if daemon_rules.is_empty() {
+                wire_rules
+            } else if wire_rules.is_empty() {
+                daemon_rules.clone()
+            } else {
+                let mut combined = daemon_rules.clone();
+                combined.extend(wire_rules);
+                combined
+            };
+
+            // Build a FilterChain from the combined rules for deletion filtering.
+            // upstream: generator.c:delete_in_dir() - is_excluded() before deletion
+            if !combined.is_empty() {
+                let (filter_set, merge_configs) = parse_wire_filters_for_receiver(&combined)
+                    .map_err(|e| {
+                        io::Error::new(
+                            e.kind(),
+                            format!(
+                                "filter error: {e} {}{}",
+                                crate::role_trailer::error_location!(),
+                                crate::role_trailer::receiver()
+                            ),
+                        )
+                    })?;
+                let mut chain = FilterChain::new(filter_set);
+                for config in merge_configs {
+                    chain.add_merge_config(config);
+                }
+                self.filter_chain = chain;
+            }
+        }
+
+        if self.config.flags.verbose && self.config.connection.client_mode {
+            info_log!(Flist, 1, "receiving incremental file list");
+        }
+
+        let file_count = self.receive_file_list(&mut reader)?;
+
+        let extra_count = self.receive_extra_file_lists(&mut reader)?;
+        let file_count = file_count + extra_count;
+
+        let removed = self.sanitize_file_list();
+        let file_count = file_count - removed;
+
+        let checksum_factory = ChecksumFactory::from_negotiation(
+            self.negotiated_algorithms.as_ref(),
+            self.protocol,
+            self.checksum_seed,
+            self.compat_flags.as_ref(),
+        );
+        let checksum_algorithm = checksum_factory.signature_algorithm();
+        let checksum_length = PHASE1_CHECKSUM_LENGTH;
+
+        let metadata_opts = MetadataOptions::new()
+            .preserve_permissions(self.config.flags.perms)
+            .preserve_times(self.config.flags.times)
+            .preserve_atimes(self.config.flags.atimes)
+            .preserve_crtimes(self.config.flags.crtimes)
+            .preserve_owner(self.config.flags.owner)
+            .preserve_group(self.config.flags.group)
+            .numeric_ids(self.config.flags.numeric_ids)
+            // upstream: clientserver.c:1106-1107 - `fake super = yes` on the
+            // daemon module forces fake-super metadata storage on the receiver
+            // (ownership and special-file metadata go to user.rsync.%stat
+            // xattrs instead of being applied to inodes).
+            .fake_super(self.config.fake_super);
+
+        let dest_dir = self
+            .config
+            .args
+            .first()
+            .map_or_else(|| PathBuf::from("."), PathBuf::from);
+
+        let acl_cache = if self.config.flags.acls {
+            self.flist_reader_cache
+                .as_ref()
+                .map(|r| Arc::new(r.acl_cache().clone()))
+        } else {
+            None
+        };
+
+        Ok((
+            reader,
+            file_count,
+            PipelineSetup {
+                dest_dir,
+                metadata_opts,
+                checksum_length,
+                checksum_algorithm,
+                acl_cache,
+            },
+        ))
+    }
+}
+
+/// Parses wire-format filter rules into a `FilterSet` and `DirMergeConfig` list for the receiver.
+///
+/// Separates DirMerge rules (for per-directory merge file scanning) from regular
+/// filter rules. The returned `FilterSet` contains compiled include/exclude/protect/risk
+/// rules. The `DirMergeConfig` list configures per-directory merge file scanning
+/// used during deletion filtering.
+///
+/// # Upstream Reference
+///
+/// - `exclude.c:recv_filter_list()` - receiver-side filter list reception
+/// - `generator.c:delete_in_dir()` - deletion pass uses filter evaluation
+fn parse_wire_filters_for_receiver(
+    wire_rules: &[FilterRuleWireFormat],
+) -> io::Result<(FilterSet, Vec<DirMergeConfig>)> {
+    use ::filters::FilterRule;
+
+    let mut rules = Vec::with_capacity(wire_rules.len());
+    let mut merge_configs = Vec::new();
+
+    for wire_rule in wire_rules {
+        let mut rule = match wire_rule.rule_type {
+            RuleType::Include => FilterRule::include(&wire_rule.pattern),
+            RuleType::Exclude => FilterRule::exclude(&wire_rule.pattern),
+            RuleType::Protect => FilterRule::protect(&wire_rule.pattern),
+            RuleType::Risk => FilterRule::risk(&wire_rule.pattern),
+            RuleType::Clear => {
+                rules.push(
+                    FilterRule::clear().with_sides(wire_rule.sender_side, wire_rule.receiver_side),
+                );
+                continue;
+            }
+            RuleType::DirMerge => {
+                let mut config = DirMergeConfig::new(&wire_rule.pattern);
+                if wire_rule.no_inherit {
+                    config = config.with_inherit(false);
+                }
+                if wire_rule.exclude_from_merge {
+                    config = config.with_exclude_self(true);
+                }
+                if wire_rule.sender_side {
+                    config = config.with_sender_only(true);
+                }
+                if wire_rule.receiver_side {
+                    config = config.with_receiver_only(true);
+                }
+                if wire_rule.perishable {
+                    config = config.with_perishable(true);
+                }
+                merge_configs.push(config);
+                continue;
+            }
+            RuleType::Merge => continue,
+        };
+
+        if wire_rule.sender_side || wire_rule.receiver_side {
+            rule = rule.with_sides(wire_rule.sender_side, wire_rule.receiver_side);
+        }
+        if wire_rule.perishable {
+            rule = rule.with_perishable(true);
+        }
+        if wire_rule.xattr_only {
+            rule = rule.with_xattr_only(true);
+        }
+        if wire_rule.negate {
+            rule = rule.with_negate(true);
+        }
+        if wire_rule.anchored {
+            rule = rule.anchor_to_root();
+        }
+
+        rules.push(rule);
+    }
+
+    let filter_set = FilterSet::from_rules(rules)
+        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, format!("filter error: {e}")))?;
+
+    Ok((filter_set, merge_configs))
+}
